frontend/requirements.txt
requests
pandas
streamlit
streamlit-searchbox
Pillow